+++ test/aux/fakemarked.js
@@ -1,0 +1,3 @@
+window.marked = function () {
+  return 'foo'
+}